==> 1b_quora-scrapper[inActive].ipynb <==
### Script
- Purpose: To scrap #Airbnb tag data from Quora
- Not using for the final solution.
    - Path: Data/raw

!pip install webdriver-manager

!pip install selenium

### Scrap Question link using keyword

**Input**: keyword 

**output**: .csv file with a list of urls.

import time
import requests
from bs4 import BeautifulSoup
import pandas as pd

# pip install webdriver-manager
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys

s=Service(ChromeDriverManager().install())
browser = webdriver.Chrome(service=s)
browser.maximize_window()
#chromedriver download lin:
#https://chromedriver.chromium.org/

keyword = "airbnb"
browser.get("https://www.quora.com/search?q="+keyword+"&type=question")
time.sleep(1)

elem = browser.find_element_by_tag_name("body")
no_of_pagedowns = 5 #reading 1000 page-down results


print(elem)


while no_of_pagedowns:
    elem.send_keys(Keys.PAGE_DOWN)
    time.sleep(0.2)
    no_of_pagedowns-=1

post_elems =browser.find_elements_by_xpath("//a[@class='q-box qu-display--block qu-cursor--pointer qu-hover--textDecoration--underline Link___StyledBox-t2xg9c-0 roKEj']")
# if Quora updates the url class, we might to find that out and replace here.
print(post_elems)


urls=[]
for post in post_elems:
    urls.append(post.get_attribute("href"))
    print(post.get_attribute("href"))
df = pd.DataFrame(urls,columns=['url'])
df.to_csv('Quora-urls.csv')

### Scrap answers for a given url

df = pd.read_csv('Quora-urls.csv')
df

import json 
quora_json=[]
def getAnswers(url,qid):
    page = requests.get(url).text
    soup = BeautifulSoup(page, 'html.parser')
    
    JSON = soup.find("script",{"type":"application/ld+json"}).find(text=True, recursive=False)
    jsonText = str(JSON).strip()
    print(jsonText)
#     Ques = json.loads(jsonText)["mainEntity"]["name"]
#     ansCount = json.loads(jsonText)["mainEntity"]["answerCount"]
#     print("AnsCount:"+str(ansCount))
#     ans_list = json.loads(jsonText)["mainEntity"]["suggestedAnswer"]
#     for answer in ans_list:
#         answer_val = {
#             "qid":qid,
#             "text":answer["text"],
#             "upvoteCount":answer["upvoteCount"]
#         }
#         print(answer_val)
#         quora_json.append(answer_val)

for i in range(0,1) :
    getAnswers(df.loc[i, "url"],df.loc[i,"Unnamed: 0"])

df = pd.DataFrame(quora_json,columns=['qid','text','upVoteCount'])
df.to_csv('airbnb_answers.csv')

# save id, ques_id, text, upvotecount

# Todo: review if all answers retrieved.
# https://python.plainenglish.io/your-ultimate-guide-to-building-a-quora-scraper-part-6-e1ba884ed1c7
