==> 2a_Dialog_Gen.ipynb <==
## Purpose
- **Input:** Airbnb - listings, reviews, single location.
- **Output:** Creating the below files to generate dialogues, train data for conversation recommendation module.
    - listings_info_filter
    - ratings_filter
    - listings_slot_value_filter
    
- 16k listings | 397k reviews

### Import libraries, raw data

import pandas as pd
import os
import numpy as np

ls

root='./Data/raw/'
processor = './Data/processing/'

listings = pd.read_csv(root+'listings.csv.gz')

## These are the columns that will be used in the dialog generation.
listingTemplate = pd.read_csv(processor+'listing_Template.csv')

### Compute unique entities, get sentiment score(user rating 0-5) for user comments.

listings = listings[listingTemplate.columns]
listings.dropna()
listings.info()

listings = listings.rename(columns={"neighbourhood": "location"})
listings.info()

len(listings)

#### Extract unique entities

# https://towardsdatascience.com/dealing-with-list-values-in-pandas-dataframes-a177e534f173
def to_1D(series):
    return pd.Series([x for _list in series for x in _list])

listings["amenities"] = listings["amenities"].apply(eval)
amenities = to_1D(listings["amenities"]).unique()

#[_id,_labels,name]
_id=[]
_labels=[]
name=[]
_start = []
_end = []
_rating = []
_relationship=[]

def convertList2Dict(lst,prev_index,label):
    dict_ret = dict() 
    for index,value in enumerate(lst):
        rec_index = index + prev_index
        dict_ret[value] = rec_index

        ##
        _id.append(rec_index)
        _labels.append(label)
        name.append(value)
        _start.append(np.NaN)
        _end.append(np.NaN)
        _relationship.append(np.NaN)
        _rating.append(np.NaN)

    prev_index = rec_index
    return dict_ret, prev_index

location = listings["location"].dropna().str.split(",", n = 2, expand = True)

listings["City"]= location[0]
listings["State"]= location[1]
listings["Country"]= location[2]

City = listings["City"].dropna().unique()
State = listings["State"].dropna().unique()
Country = listings["Country"].dropna().unique()

listings["price"].min()
listings["price"].max()
listings["price"] = listings["price"].replace('[\$,]', '', regex=True).astype(float)
listings["price"] = pd.cut(listings['price'], [0, 250, 500,750,999], labels=['low', 'average', 'high','expensive'])
price = listings["price"].dropna().unique()

listings["review_scores_rating"].dropna()
listings["review_scores_rating"] = pd.cut(listings['review_scores_rating'], [0, 1, 2, 3,4,5], labels=['horrible', 'bad', 'average','good','very good'])
rating = listings["review_scores_rating"].unique()

listings["review_scores_accuracy"].dropna()
listings["review_scores_accuracy"] = pd.cut(listings['review_scores_accuracy'], [0, 1, 2, 3,4,5], labels=['horrible', 'bad', 'average','good','very good'])
acc_rating = listings["review_scores_accuracy"].unique()

listings["review_scores_cleanliness"].dropna()
listings["review_scores_cleanliness"] = pd.cut(listings['review_scores_cleanliness'], [0, 1, 2, 3,4,5], labels=['horrible', 'bad', 'average','good','very good'])
clean_rating = listings["review_scores_cleanliness"].unique()

listings["review_scores_checkin"].dropna()
listings["review_scores_checkin"] = pd.cut(listings['review_scores_checkin'], [0, 1, 2, 3,4,5], labels=['horrible', 'bad', 'average','good','smooth'])
checkin_rating = listings["review_scores_checkin"].unique()

listings["review_scores_communication"].dropna()
listings["review_scores_communication"] = pd.cut(listings['review_scores_communication'], [0, 1, 2, 3,4,5], labels=['horrible', 'bad', 'average','good','very good'])
communication_rating = listings["review_scores_communication"].unique()

listings["review_scores_location"].dropna()
listings["review_scores_location"] = pd.cut(listings['review_scores_location'], [0, 1, 2, 3,4,5], labels=['horrible', 'bad', 'average','good','very good'])
location_rating = listings["review_scores_location"].unique()

host_identity_verified = listings["host_identity_verified"].dropna().unique()    
property_type = listings["property_type"].dropna().unique()    
room_type = listings["room_type"].dropna().unique()    
accommodates = listings["accommodates"].dropna().unique().astype(float)    
bathrooms = listings["bathrooms_text"].dropna().unique()    
bedrooms = listings["bedrooms"].dropna().unique()    
beds = listings["beds"].dropna().unique()    

accommodates=np.delete(accommodates,np.where(accommodates == 0))

amenities_dict, last_index = convertList2Dict(amenities,0,':Amenities')
City_dict, last_index = convertList2Dict(City,(last_index+1),':City')
State_dict, last_index = convertList2Dict(State,(last_index+1),':State')
Country_dict, last_index = convertList2Dict(Country,(last_index+1),':Country')
price_dict, last_index = convertList2Dict(price,(last_index+1),':price')
rating_dict, last_index = convertList2Dict(rating,(last_index+1),':rating')
acc_rating_dict, last_index = convertList2Dict(acc_rating,(last_index+1),':acc_rating')
clean_rating_dict, last_index = convertList2Dict(clean_rating,(last_index+1),':clean_rating')
checkin_rating_dict, last_index = convertList2Dict(checkin_rating,(last_index+1),':checkin_rating')
communication_rating_dict, last_index = convertList2Dict(communication_rating,(last_index+1),':communication_rating')
location_rating_dict, last_index = convertList2Dict(location_rating,(last_index+1),':location_rating')
host_identity_verified_dict, last_index = convertList2Dict(host_identity_verified,(last_index+1),':host_identity_verified')
property_type_dict, last_index = convertList2Dict(property_type,(last_index+1),':property_type')
room_type_dict, last_index = convertList2Dict(room_type,(last_index+1),':room_type')
accommodates_dict, last_index = convertList2Dict(accommodates,(last_index+1),':accommodates')
bathrooms_dict, last_index = convertList2Dict(bathrooms,(last_index+1),':bathrooms')
bedrooms_dict, last_index = convertList2Dict(bedrooms,(last_index+1),':bedrooms')
beds_dict, last_index = convertList2Dict(beds,(last_index+1),':beds')

import json 
dictionary = {
 "amenities":amenities_dict,
 "City": City_dict,
 "State":State_dict,
 "Country":Country_dict,
 "price":price_dict,
 "rating":rating_dict,
 "accuracy":acc_rating_dict,
 "cleanliness":clean_rating_dict,
 "checkin":checkin_rating_dict,
 "communication":communication_rating_dict,
 "location":location_rating_dict,
 "host_identity_verified":host_identity_verified_dict,
 "property_type":property_type_dict,
 "room_type":room_type_dict,
 "accommodates":accommodates_dict, 
 "bathrooms":bathrooms_dict,
 "bedrooms":bedrooms_dict,
 "beds":beds_dict
}
with open(processor+'Processed_Airbnb/listings_entities_filter.json', 'w', encoding='utf-8') as f:
    f.write(json.dumps(dictionary, ensure_ascii=False))

df = pd.read_json(processor+'Processed_Airbnb/listings_entities_filter.json')
df.to_csv(processor+'Processed_Airbnb/listings_entities_filter.csv')

with open(processor+'Processed_Airbnb/listings_info_filter.json', 'w', encoding='utf-8') as f:
    f.write(listings.to_json(orient = 'records'))

reviews0 = pd.read_csv(processor+'Processed_Airbnb/ratings_filter.csv')
reviews0.info()

reviews = reviews0[['listing_id','id','reviewer_id','comments','rating']]

reviews.head(1)

## Create Neo4J input file


reviewer = reviews['reviewer_id'].unique() 
for user in reviewer:
    if user>1:
        last_index = last_index+1
        _id.append(last_index)
        _labels.append(':User')
        name.append(int(user))
        _start.append(np.NaN)
        _end.append(np.NaN)
        _relationship.append(np.NaN)
        _rating.append(np.NaN)

listing = listings['id'].unique() 
for list in listing:
    if(list>1):
        last_index = last_index+1
        _id.append(last_index)
        _labels.append(':Listing')
        name.append(int(list))
        _start.append(np.NaN)
        _end.append(np.NaN)
        _relationship.append(np.NaN)
        _rating.append(np.NaN)

reviews.head(2)

#form a list of (u_id,l_id)

data = {'_id':_id,
        '_labels':_labels,
        'name':name,
        '_start':_start,
        '_end':_end,
        '_type':_relationship,
        'rating':_rating
        }

#Create DataFrame
neo4J_format_df = pd.DataFrame(data)
neo4J_format_df.to_csv(processor+'neo4J_part1.csv',index=False)

# for every review, get id of listing and user from the neo4j csv and create a rating edge.
import csv

def find_nodeVal(type, name, filename):
   with open(filename) as f:
     reader = csv.reader(f, delimiter=',')
     for row in reader:
         if(row[1]==type and row[2]==name):
            return row
fr = find_nodeVal(':User','10952',processor+'neo4J_part1.csv')
fr = find_nodeVal(':Listing','2818',processor+'neo4J_part1.csv')

def find_edge(l_id, u_id, filename):
   with open(filename) as f:
     reader = csv.reader(f, delimiter=',')
     for row in reader:
         if(row[0]==l_id):
            print(row)
            return row

fr = find_edge('2818','10952',processor+'Processed_Airbnb/ratings_filter.csv')


for index in reviews.index:
    user_id = int(reviews['reviewer_id'][index])
    listing_id = int(reviews['listing_id'][index])
    rating = int(reviews['rating'][index])
    if rating>=1:
        if(index==5):
            break
        g_reviewer_id = find_nodeVal(':User',str(user_id),processor+'neo4J_part1.csv')
        g_listing_id = find_nodeVal(':Listing',str(listing_id),processor+'neo4J_part1.csv')
        _id.append(np.NaN)
        _labels.append(np.NaN)
        name.append(np.NaN)
        _start.append(int(g_reviewer_id[0]))
        _end.append(int(g_listing_id[0]))
        _rating.append(int(rating))
        _relationship.append('rated') #user rated listing => rating


data = {'_id':_id,
        '_labels':_labels,
        'name':name,
        '_start':_start,
        '_end':_end,
        '_type':_relationship,
        'rating':_rating
        }
# Create DataFrame
neo4J_format_df = pd.DataFrame(data)
neo4J_format_df.to_csv(processor+'neo4J_part2.csv',index=False)

#### Compute rating from review : StandfordNLP Sentiment Analysis

# !pip install pycorenlp
# https://towardsdatascience.com/natural-language-processing-using-stanfords-corenlp-d9e64c1e1024

from pycorenlp import StanfordCoreNLP
# Need to start the java server.stanford-corenlp-4.3.1 % java -mx4g -cp "*" edu.stanford.nlp.pipeline.StanfordCoreNLPServer -port 9000 -timeout 50000
nlp_wrapper = StanfordCoreNLP('http://localhost:9000')

def computeRatingfromReview(doc):
    overall_sentiment=0
    count=0
    
    annot_doc = nlp_wrapper.annotate(doc,
    properties={
       'annotators': 'sentiment',
       'outputFormat': 'json',
       'timeout': 100000,
    })
    
    try:
        for s in annot_doc["sentences"]:
            sentiment = s["sentiment"]
            if(sentiment=='Verypositive'):
                overall_sentiment = overall_sentiment+5
            elif(sentiment=='Positive'):
                overall_sentiment = overall_sentiment+4
            elif(sentiment=='Neutral'):
                overall_sentiment = overall_sentiment+3
            elif(sentiment=='Negative'):
                overall_sentiment = overall_sentiment+2
            elif(sentiment=='Verynegative'):
                overall_sentiment = overall_sentiment+1
            count+=1

    except:
            return 0
        
    return ((int)(overall_sentiment/count))

len(reviews)

# reviews=reviews[:100]
import time
st = time.time()
print(st)

reviews['rating'] = reviews['comments']

for x in range(len(reviews.index)):
    try:
        reviews['rating'].iloc[x] = computeRatingfromReview(reviews['comments'].iloc[x])
    except:
        reviews['rating'].iloc[x] = 0

reviews = reviews[['listing_id','reviewer_id','rating']]
reviews = reviews.rename(columns={"listing_id": "business_id", "reviewer_id": "user_id","rating":"stars"})

reviews.head()

reviews.to_csv(processor+'Processed_Airbnb/ratings_filter.csv',index=False)

end = time.time()
print(end)

review_rated = pd.read_csv(processor+'Processed_Airbnb/ratings_filter.csv')
reviews = pd.read_csv(root+'reviews.csv.gz')
reviews['rating'] = review_rated['rating']
reviews = reviews.dropna()
reviews.to_csv(processor+'Processed_Airbnb/ratings_filter.csv',index=False)

reviews
