==> 2c_ProcessText.ipynb <==
### Script
- **Input:** Airbnb - listings, reviews, single location.
- **Output:** Creating the below files to generate dialogues, train data for RASA and backend data for Neo4j.
    
    - listings_info_filter
    - ratings_filter
    - listings_slot_value_filter
    - Neo4j_nodes.csv
    
- Data Size: 5.5k Airbnb listings | 250k user reviews

### References:

- https://neo4j.com/developer/guide-import-csv/

### Import libraries, raw data

import pandas as pd
import os
import numpy as np
import warnings
warnings.filterwarnings('ignore')

### Preprocessing helper methods

root='./Data/raw/'
processor = './Data/processing/'

listings = pd.read_csv(root+'listings.csv.gz')


## These are the columns that will be used in the dialog generation.
listingTemplate = pd.read_csv(processor+'listing_Template.csv')
listing_reviews = pd.read_csv(processor+'Processed_Airbnb/listing_with_reviews.csv')
listing_reviews = listing_reviews.set_index('listing_id').T.to_dict('list')

### Compute unique entities, create Neo4j nodes & edge map.

# listings = listings[listingTemplate.columns]
listings.dropna()
listings.info()

len(listings)

### Add Geo-location, check more efficient way to add location.
listings['City'] = 'Amsterdam'
listings['Country'] = 'Netherlands'
listings[listings['neighbourhood_cleansed'].isna()]
listings['neighbourhood'] = listings['neighbourhood'].fillna(listings['neighbourhood_cleansed'])
listings['neighbourhood'] = listings['neighbourhood'].str.strip()
listings['State'] = listings['neighbourhood']

#### Add all unique entities

### Add listings text description
# 

col_parse = ['id',
 'picture_url',
 'host_identity_verified',
 'neighbourhood',
 'neighbourhood_cleansed',
 'property_type',
 'room_type',
 'accommodates',
 'bathrooms_text',
 'bedrooms',
 'beds',
 'price',
 'City',
 'Country',
 'State']

listings['Listing_Text'] = listings['City']

col = list(listings.columns.values)
for i,j in listings.iterrows():
    for c_name in col_parse:
        listings['Listing_Text'][i] = listings['Listing_Text'][i] + str(c_name) + ':' + str(j[str(c_name)])+';'
    


listings['Review_Text'] = listings['Listing_Text']
for i,j in listings.iterrows():
    lid = (listings['id'][i])
    try: 
        listings['Review_Text'][i] = listing_reviews[lid][1]
    except:
        listings['Review_Text'][i] = 'no review found'

listings['Review_Text']

listings['Review_Text']= listings['Review_Text'].apply(lambda x: x.lower())
listings['Listing_Text']= listings['Listing_Text'].apply(lambda x: x.lower())


listings['Review_Text']

### Save Processed listing

listings.to_csv(processor+'Processed_Airbnb/listings_text_processed.csv')
