==> 4a_GenerateRASA_Conversational_Data.ipynb <==
### Script
- **Input:** FAQ generated in 2b and listing entities generated in 2d.
- **Output:** auto-generated nlu.yml for RASA solution.
    
- Approx 300K intents

root = './RASA_ConceptNet5/data/'
processed = './Data/processing/Processed_Airbnb/'
test = './Data/test'

### Import

import pandas as pd 

## Generate nlu.yml

### Gather FAQs

FAQ1 = pd.read_csv(processed+'FAQ_1.json')
FAQ2 = pd.read_csv(processed+'FAQ_2.json')
FAQ3 = pd.read_csv(processed+'FAQ_3.json')


FAQ =  pd.concat([FAQ1, FAQ2, FAQ3], axis=0)

FAQ.head(1)

### Add intent type

FAQ['intent'] = 'affirm_FAQ'
FAQ['type'] = 'intent'

FAQ.head(5)

### Append entity and entity value to the intent.

#Remove ? from the question 
#Append entity value and entity at the end and append '?' => [answer](wordid)?

FAQ['question'] = FAQ['question'].str.replace(r'?', ' ',regex=True)
FAQ.head(5)

# FAQ['question'] = FAQ['question'].astype(str) +'['+FAQ['answer'].astype(str)+'](wordid)?'
FAQ['question'] = FAQ['question'].astype(str)

FAQ.head(5)

### Auto-generate NLU intents

nlus = FAQ.groupby('intent')['question'].apply(list).to_dict()

answer = FAQ["answer"].unique()
nlu=[]
query=[]
entity=[]
for val in answer:
    questionString1="find me close matches for a place near "
    questionString2="get me recommendations with features close to"
    questionString3="I am looking for a place with facilities similar to "
    questionString4="get me recommendations with facilities similar to "
    questionString5="find similar recommendations like a "
    questionString6="find places with amenities similar to "
    questionString7="find places suitable for "
     

    n = val
    nlu.append(questionString1+"["+str(n)+"](wordid)")
    nlu.append(questionString2+"["+str(n)+"](wordid)")
    nlu.append(questionString3+"["+str(n)+"](wordid)")
    nlu.append(questionString4+"["+str(n)+"](wordid)")
    nlu.append(questionString5+"["+str(n)+"](wordid)")
    nlu.append(questionString6+"["+str(n)+"](wordid)")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))
    query.append(questionString3+" "+str(n))
    query.append(questionString4+" "+str(n))
    query.append(questionString5+" "+str(n))
    query.append(questionString6+" "+str(n))
    query.append(questionString7+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))

# #;TYPE;INTENT;VALUE
# f = open(root+"nlu.yml","a")

# for intent in nlus:
#     f.write("\n- intent: {}\n".format(intent))
#     f.write("  examples: |\n")
#     for value in nlus[intent]:
#         f.write("    - {}\n".format(value))

# f.write("\n- intent: {}\n".format("affirm_neo4j_conceptNet5_review"))
# f.write("  examples: |\n")
# for value in nlu:
#     f.write("    - {}\n".format(value.lower()))
# f.close()



### Generate listings nlu

import csv 

listing_reviews = pd.read_csv(processed+'listings_text_processed.csv')

listing_reviews = listing_reviews[['neighbourhood','neighbourhood_cleansed','property_type'
                                  ,'room_type','accommodates','bathrooms','bathrooms_text'
                                  ,'bedrooms','beds','amenities','price']]


neighborhood =  listing_reviews['neighbourhood'].unique()
loc =  listing_reviews['neighbourhood'].unique()
accomodates = listing_reviews['accommodates'].unique()
bathrooms = listing_reviews['bathrooms'].unique()
bedrooms = listing_reviews['bedrooms'].unique()
beds = listing_reviews['beds'].unique()
price = listing_reviews['price'].unique()
bathrooms_text = listing_reviews['bathrooms_text'].unique()
property_type = listing_reviews['property_type'].unique()
room_type = listing_reviews['room_type'].unique()
amenities = listing_reviews['amenities'].unique()

nlu = []
for val in loc:
    questionString1="find me a place near "
    questionString2="get me recommendations with "
    questionString3="I am looking for a place with facilities like "
    questionString4="get me recommendations with facilities to " 

    if ',' in val:
        neigh = val.split(',')
        for n in neigh:
            nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
            nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")
            nlu.append(questionString3+"["+n.strip('\"').strip()+"](wordid)")
            nlu.append(questionString4+"["+n.strip('\"').strip()+"](wordid)")

    else:
        n = val
        nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
        nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")
        nlu.append(questionString3+"["+n.strip('\"').strip()+"](wordid)")
        nlu.append(questionString4+"["+n.strip('\"').strip()+"](wordid)")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))
    query.append(questionString3+" "+str(n))
    query.append(questionString4+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))
    entity.append(str(n))

for val in neighborhood:
    questionString1="find me a place in the neighborhood of "
    questionString2="find me a place in "

    if ',' in val:
        neigh = val.split(',')
        for n in neigh:
            nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
            nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")
    else:
        n = val
        nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
        nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")
    
    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in accomodates:
    questionString1="find me a place that can accomodate "
    questionString2="find me a place for "
    
    n = str(val)
    nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
    nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in bathrooms:
    questionString1="find me a place with "
    questionString2="looking for a place with "

    n = str(val)
    nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) bathrooms")
    nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) bathrooms")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in bedrooms:
    questionString1="find me a place with "
    questionString2="looking for a place with "

    n = str(val)
    nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) bedrooms")
    nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) bedrooms")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in beds:
    questionString1="find me a place with "
    questionString2="looking for a place with "

    n = str(val)
    nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) beds")
    nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) beds")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in price:
    questionString1="find me a place in the range of "
    questionString2="looking for a place within the price "

    n = val
    nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
    nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in bathrooms_text:
    questionString1="find me a place with bath that is "
    questionString2="looking for a place with bath features like "

    n = str(val)
    nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
    nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in amenities:
    questionString1="find me a place with amenities like "
    questionString2="I am looking for a place with amenities like "
    val = val.strip('[').strip(']')
    if ',' in val:
        neigh = val.split(',')
        for n in neigh:
            nlu.append(questionString1+"["+n.replace('"', '').strip()+"](wordid)")
            nlu.append(questionString2+"["+n.replace('"', '').strip()+"](wordid)")
    else:
        n = val
        nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid)")
        nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid)")
    
    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in property_type:
    questionString1="find me a "
    questionString2="looking for a "

    if ',' in val:
        neigh = val.split(',')
        for n in neigh:
            nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) property")
            nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) property")
    else:
        n = val
        nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) property")
        nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) property")
     
    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))

for val in room_type:
    questionString1="find me a "
    questionString2="looking for a"

    if ',' in val:
        neigh = val.split(',')
        for n in neigh:
            nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) room")
            nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) room")
    else:
        n = val
        nlu.append(questionString1+"["+n.strip('\"').strip()+"](wordid) room")
        nlu.append(questionString2+"["+n.strip('\"').strip()+"](wordid) room")

    query.append(questionString1+" "+str(n))
    query.append(questionString2+" "+str(n))

    entity.append(str(n))
    entity.append(str(n))



# #;TYPE;INTENT;VALUE
# f = open(root+"nlu.yml","a")

# f.write("\n- intent: {}\n".format("affirm_neo4j_conceptNet5_listing"))
# f.write("  examples: |\n")
# for value in nlu:
#     f.write("    - {}\n".format(value.lower()))
# f.close()

df = pd.DataFrame(list(zip(query, entity)),
               columns =['user_query', 'entity'])
df

df.to_csv(test+'/test_user_queries.csv')
